Add functionality for editing and updating your own profile.

+++ app/Http/Controllers/UserController.php
@@ -13,6 +13,9 @@ class UserController extends Controller
      */
     public function index()
     {
+        // This is the UserPolicy at work...
+        // Note to self: check UserPolicy.php for the corresponding methods.
+        $this->authorize('viewAny', User::class);
         $users = User::all();
         return view('users.index', compact('users'));
     }
@@ -22,6 +25,7 @@ class UserController extends Controller
      */
     public function create()
     {
+        $this->authorize('create', User::class);
         return view('users.create');
     }
 
@@ -30,6 +34,7 @@ class UserController extends Controller
      */
     public function store(Request $request)
     {
+        $this->authorize('create', User::class);
         $incomingFields = $request->validate([
             'first_name' => 'required',
             'last_name' => 'required',
@@ -47,9 +52,9 @@ class UserController extends Controller
     /**
      * Display the specified resource.
      */
-    public function show(string $id)
+    public function show(User $user)
     {
-        //
+        $this->authorize('view', User::class);
     }
 
     /**
@@ -58,7 +63,21 @@ class UserController extends Controller
     public function edit(User $user)
     {
         $this->authorize('update', $user);
-        return view('users.edit', compact('user'));
+        return view('users.edit', compact('user'),
+            // I'm using this property as a way to determine how a specific element is rendered in the blade template
+            // This is also used in the profile method below,
+            // but in that case I just hard coded the boolean value to be true.
+            //
+            // Since the users.edit view is being shared for Super Admins editing anyone's info, and for
+            // any user editing their own info, there needs to be an 'isOwnProfile' property for both methods.
+            // Since a Super Admin could access their own profile from the users index, they can basically access
+            // the users.edit view from two different controller methods, so this property needs to be present but
+            // must work dynamically so the page renders appropriately.
+            [
+                'isOwnProfile' => auth()->id() === $user->id,
+                'formAction' => route("users.update", $user),
+
+                ]);
     }
 
     /**
@@ -103,7 +122,44 @@ class UserController extends Controller
      */
     public function destroy(User $user)
     {
+        $this->authorize('delete', $user);
         $user->delete();
         return redirect()->route('users.index')->with('success', 'User deleted successfully');
+    }
+
+
+    public function editProfile(){
+        return view('users.edit', [
+            'user' => auth()->user(),
+            'isOwnProfile' => true,
+            'formAction' => route('profile.update'),
+        ]);
+    }
+
+    public function updateProfile(Request $request){
+        $user = auth()->user();
+        $incomingFields = $request->validate([
+            'first_name' => 'required',
+            'last_name' => 'required',
+            'email' => [
+                'required',
+                'email',
+                Rule::unique('users', 'email')->ignore($user),
+            ],
+            'password' => 'nullable|confirmed',
+        ]);
+
+        if (!empty($incomingFields['password'])) {
+            // If the password field is filled, the password will be hashed and updated.
+            $incomingFields['password'] = bcrypt($incomingFields['password']);
+        } else {
+            // If the password field is empty, it is dropped from the $incomingFields array and the original password is kept.
+            unset($incomingFields['password']);
+        }
+
+        $user->update($incomingFields);
+
+        return redirect()->route('home')->with('success', 'Profile updated successfully');
+
     }
 }
